Move small components to components/lib with index.js
