+++ components/lib/Overlay.js
@@ -1,0 +1,10 @@
+const Overlay = ({ close, children }) => (
+  <div
+    className="z-20 bg-black bg-opacity-50 px-2 pt-10 fixed top-0 left-0 right-0 bottom-0"
+    onClick={close}
+  >
+    {children}
+  </div>
+)
+
+export default Overlay
