+++ pages/posts/[slug]/index.js
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import Image from 'next/image'
 import Layout from '../../../components/Layout'
-import DateSpan from '../../../components/DateSpan'
+import { DateSpan } from '../../../components/lib'
 import { PostArticle } from '../../../components/Post'
 import { useSession } from '../../../lib/auth'
 import { fetchPostList, fetchOnePost } from '../../../lib/posts'
